requirements.txt
# MentorBoxAI - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.0.0

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP (used elsewhere, keep for compatibility)
requests>=2.31.0

# AWS SDK — Bedrock (Nova 2 Lite) + S3/DynamoDB
boto3>=1.34.0
botocore>=1.34.0

# Animation
manim>=0.18.0

# Utilities
numpy>=1.24.0
black>=24.0.0
